+++ csrc/ops/arith.h
@@ -98,7 +98,7 @@ NVF_API TensorView* binaryOp(
 // of size-1 dimension is automatically converted to squeeze.
 NVF_API TensorView* reductionOp(
     BinaryOpType reduction_op_type,
-    const std::vector<int>& axes,
+    const std::vector<int64_t>& axes,
     Val* init,
     TensorView* v1,
     bool keep_dim = false,
@@ -108,7 +108,7 @@ NVF_API TensorView* reductionOp(
 // reduction into squeeze and don't convert size-0 reduction into full.
 NVF_API TensorView* reductionOpRaw(
     BinaryOpType reduction_op_type,
-    const std::vector<int>& axes,
+    const std::vector<int64_t>& axes,
     Val* init,
     TensorView* v1,
     bool keep_dim = false,
@@ -134,7 +134,7 @@ class WelfordResult {
 //! ops are added.
 NVF_API WelfordResult Welford(
     TensorView* tv,
-    const std::vector<int>& axes,
+    const std::vector<int64_t>& axes,
     TensorView* init_avg = nullptr,
     TensorView* init_var = nullptr,
     // Initializes to 0 in function definition, doing this so we don't have to
@@ -145,7 +145,7 @@ NVF_API WelfordResult Welford(
 //! squeeze/full.
 WelfordResult WelfordRaw(
     TensorView* tv,
-    const std::vector<int>& axes,
+    const std::vector<int64_t>& axes,
     TensorView* init_avg = nullptr,
     TensorView* init_var = nullptr,
     // Initializes to 0 in function definition, doing this so we don't have to
@@ -587,25 +587,25 @@ TensorView* complex(TensorView* v1, TensorView* v2);
 // REDUCTION OPERATIONS
 NVF_API TensorView* sum(
     TensorView* v1,
-    const std::vector<int>& reduction_axes,
+    const std::vector<int64_t>& reduction_axes,
     bool keep_dim = false,
     DataType dtype = DataType::Null);
 
 NVF_API TensorView* prod(
     TensorView* v1,
-    const std::vector<int>& reduction_axes,
+    const std::vector<int64_t>& reduction_axes,
     bool keep_dim = false,
     DataType dtype = DataType::Null);
 
 NVF_API TensorView* max(
     TensorView* v1,
-    const std::vector<int>& reduction_axes,
+    const std::vector<int64_t>& reduction_axes,
     bool keep_dim = false,
     DataType dtype = DataType::Null);
 
 NVF_API TensorView* min(
     TensorView* v1,
-    const std::vector<int>& reduction_axes,
+    const std::vector<int64_t>& reduction_axes,
     bool keep_dim = false,
     DataType dtype = DataType::Null);
 
@@ -794,7 +794,7 @@ TensorView* viewAsScalar(TensorView* inp);
 NVF_API TensorView* fusedMultiplySum(
     TensorView* tv_a,
     TensorView* tv_b,
-    const std::vector<int>& axes,
+    const std::vector<int64_t>& axes,
     Val* init = nullptr);
 
 // Create a tensor view from the given value. The given value can be a single
