+++ csrc/python_frontend/python_bindings.cpp
@@ -1919,7 +1919,7 @@ void initNvFuserPythonBindings(PyObject* module) {
             self.validUse(), "Attempting to add to a completed definition!"); \
         FusionDefinition* fd = self.fusion_definition;                        \
         size_t ndims = 0;                                                     \
-        std::vector<int> dims(arg.dims);                                      \
+        std::vector<int64_t> dims(arg.dims);                                  \
         std::iota(dims.begin(), dims.end(), 0);                               \
         Tensor output = fd->defineTensor(ndims);                              \
         fd->defineRecord(new ReductionOpRecord(                               \
@@ -1928,7 +1928,7 @@ void initNvFuserPythonBindings(PyObject* module) {
             ("ops." op_str),                                                  \
             record_type,                                                      \
             static_cast<TensorView* (*)(TensorView*,                          \
-                                        const std::vector<int>&,              \
+                                        const std::vector<int64_t>&,          \
                                         bool,                                 \
                                         DataType)>(op_name),                  \
             dims,                                                             \
@@ -1975,7 +1975,7 @@ void initNvFuserPythonBindings(PyObject* module) {
       op_str,                                                                 \
       [](FusionDefinition::Operators& self,                                   \
          Tensor arg,                                                          \
-         const std::vector<int>& dims,                                        \
+         const std::vector<int64_t>& dims,                                    \
          bool keepdim,                                                        \
          PrimDataType dtype) -> Tensor {                                      \
         FUSER_PERF_SCOPE("Operators." op_str);                                \
@@ -2676,7 +2676,7 @@ void initNvFuserPythonBindings(PyObject* module) {
       "var",
       [](FusionDefinition::Operators& self,
          Tensor arg,
-         std::vector<int>& dims,
+         std::vector<int64_t>& dims,
          int64_t correction,
          bool keepdim) -> Tensor {
         FUSER_PERF_SCOPE("Operators.var");
@@ -2702,7 +2702,7 @@ void initNvFuserPythonBindings(PyObject* module) {
       "var_mean",
       [](FusionDefinition::Operators& self,
          Tensor arg,
-         std::vector<int>& dims,
+         std::vector<int64_t>& dims,
          int64_t correction,
          bool keepdim) -> decltype(auto) {
         FUSER_PERF_SCOPE("Operators.var_mean");
@@ -2790,7 +2790,7 @@ void initNvFuserPythonBindings(PyObject* module) {
       py::arg("dim"));
   auto reduction_factor_func = [](FusionDefinition::SchedOperators& self,
                                   Tensor arg,
-                                  const std::vector<int>& dims) -> Tensor {
+                                  const std::vector<int64_t>& dims) -> Tensor {
     FUSER_PERF_SCOPE("SchedOperators.reduction_factor");
     NVF_CHECK(
         self.validUse(),
